ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
neo4j
transformers
torch
underthesea
google-genai
